added session timeout metric
